feat: consensus state payout methods, FullPayoutPending status, and tests
