+++ types/src/consensus_state/mod.rs
@@ -13,6 +13,7 @@ use crate::ssz_state_tree::SszStateTree;
 use crate::utils::{invalid_deposit_refund_split, parse_withdrawal_credentials};
 use crate::withdrawal::{PendingWithdrawal, WithdrawalKind, WithdrawalQueue};
 use crate::{Digest, PublicKey};
+use alloy_eips::eip4895::Withdrawal;
 use alloy_primitives::Address;
 use alloy_rpc_types_engine::ForkchoiceState;
 use bytes::{Buf, BufMut};
@@ -21,7 +22,7 @@ use commonware_cryptography::ed25519::Signature;
 use commonware_cryptography::{Verifier as _, bls12381, sha256};
 #[cfg(feature = "prom")]
 use metrics::histogram;
-use std::collections::{BTreeMap, HashSet, VecDeque};
+use std::collections::{BTreeMap, HashMap, HashSet, VecDeque};
 use std::num::NonZeroU64;
 use std::sync::Arc;
 use tracing::{error, warn};
@@ -599,15 +600,30 @@ impl ConsensusState {
     }
 
     pub fn remove_added_validator(&mut self, epoch: u64, pubkey: &PublicKey) -> bool {
-        if let Some(validators) = self.added_validators.get_mut(&epoch)
+        let removed = if let Some(validators) = self.added_validators.get_mut(&epoch)
             && let Some(pos) = validators.iter().position(|v| v.node_key == *pubkey)
         {
             validators.remove(pos);
-            self.ssz_tree
-                .rebuild_added_validators(&self.added_validators);
-            return true;
+            true
+        } else {
+            false
+        };
+        if !removed {
+            return false;
         }
-        false
+        // Drop the epoch key once its last scheduled activation is removed. An
+        // empty entry and an absent entry must not commit to different roots, so
+        // the map is kept canonical.
+        if self
+            .added_validators
+            .get(&epoch)
+            .is_some_and(|validators| validators.is_empty())
+        {
+            self.added_validators.remove(&epoch);
+        }
+        self.ssz_tree
+            .rebuild_added_validators(&self.added_validators);
+        true
     }
 
     pub fn take_pending_execution_requests(&mut self) -> Vec<alloy_primitives::Bytes> {
@@ -1506,6 +1522,120 @@ impl ConsensusState {
             .get_for_epoch_with_total_cap(epoch, max_total)
     }
 
+    /// Payout amount for one due withdrawal against `balance`, the validator's
+    /// available balance at this point in the sweep. Deposit refunds pay their
+    /// fixed amount and ignore the balance. A validator full exit (marker amount
+    /// 0) pays the entire balance. A partial pays up to its requested amount
+    /// while keeping an active validator at or above the minimum stake.
+    fn withdrawal_payout_amount(
+        entry: &PendingWithdrawal,
+        balance: u64,
+        active: bool,
+        min_stake: u64,
+    ) -> u64 {
+        match entry.kind {
+            WithdrawalKind::DepositRefund => entry.inner.amount,
+            WithdrawalKind::Validator => {
+                if entry.inner.amount == 0 {
+                    balance
+                } else {
+                    let floor = if active { min_stake } else { 0 };
+                    entry.inner.amount.min(balance.saturating_sub(floor))
+                }
+            }
+        }
+    }
+
+    /// Compute the EIP 4895 withdrawals to emit for the epoch, re clamping each
+    /// against the live balance. Read only: used at block build and verify. The
+    /// balance is debited later at apply_withdrawal_payouts. Validator exits take
+    /// priority over deposit refunds under the single per epoch total cap.
+    ///
+    /// Multiple due withdrawals for the same validator are clamped sequentially
+    /// via a transient running balance, so concurrent partials keep an active
+    /// validator at or above the minimum stake. A partial that clamps to zero is
+    /// dropped: it is not emitted here and is consumed at apply.
+    pub fn emit_withdrawal_payouts(&self, epoch: u64) -> Vec<Withdrawal> {
+        let min_stake = self.get_minimum_stake();
+        let max_total = self.get_max_withdrawals_per_epoch() as usize;
+        let mut running: HashMap<[u8; 32], u64> = HashMap::new();
+        let mut payouts = Vec::new();
+        for entry in self
+            .withdrawal_queue
+            .get_for_epoch_with_total_cap(epoch, max_total)
+        {
+            if entry.kind == WithdrawalKind::DepositRefund {
+                // Refund of a rejected deposit: the money was never in an account.
+                payouts.push(entry.inner);
+                continue;
+            }
+            let Some(account) = self.get_account(&entry.pubkey) else {
+                // Account is gone (already fully paid out). Nothing to pay.
+                continue;
+            };
+            let balance = *running.entry(entry.pubkey).or_insert(account.balance);
+            let active = account.status == ValidatorStatus::Active;
+            let payout = Self::withdrawal_payout_amount(entry, balance, active, min_stake);
+            running.insert(entry.pubkey, balance.saturating_sub(payout));
+            if payout > 0 {
+                let mut withdrawal = entry.inner;
+                withdrawal.amount = payout;
+                payouts.push(withdrawal);
+            }
+        }
+        payouts
+    }
+
+    /// Apply the withdrawal payouts that a finalized block paid out. Mirrors
+    /// emit_withdrawal_payouts against the live balance: debits each validator
+    /// payout, removes drained accounts, and consumes every processed entry
+    /// (including partials that clamp to zero). Refund payouts touch no balance.
+    /// Run at commit, once the matching block has been finalized.
+    ///
+    /// `block_withdrawals` is the EIP 4895 withdrawal list carried by the block
+    /// being committed. It must equal what this state would emit; the assert pins
+    /// the debits to exactly what the execution layer paid out, so a payload that
+    /// disagrees with the deterministic computation halts the node rather than
+    /// silently diverging the consensus balance from the execution layer.
+    /// Verification enforces the same equality before finalize, so this is
+    /// defense in depth.
+    pub fn apply_withdrawal_payouts(&mut self, epoch: u64, block_withdrawals: &[Withdrawal]) {
+        assert_eq!(
+            self.emit_withdrawal_payouts(epoch).as_slice(),
+            block_withdrawals,
+            "block withdrawals must match the payouts emitted from consensus state"
+        );
+
+        let min_stake = self.get_minimum_stake();
+        let max_total = self.get_max_withdrawals_per_epoch() as usize;
+        let indices: Vec<u64> = self
+            .withdrawal_queue
+            .get_for_epoch_with_total_cap(epoch, max_total)
+            .iter()
+            .map(|entry| entry.inner.index)
+            .collect();
+
+        for index in indices {
+            let Some(entry) = self.pop_withdrawal_by_index(epoch, index) else {
+                continue;
+            };
+            if entry.kind == WithdrawalKind::DepositRefund {
+                continue;
+            }
+            let Some(mut account) = self.get_account(&entry.pubkey).cloned() else {
+                continue;
+            };
+            let active = account.status == ValidatorStatus::Active;
+            let payout = Self::withdrawal_payout_amount(&entry, account.balance, active, min_stake);
+            account.balance = account.balance.saturating_sub(payout);
+            if account.balance == 0 {
+                self.remove_account(&entry.pubkey);
+            } else {
+                self.set_account(entry.pubkey, account);
+            }
+        }
+    }
+
     /// Get the number of pending withdrawals for a specific epoch
     pub fn get_withdrawal_count_for_epoch(&self, epoch: u64) -> usize {
         self.withdrawal_queue.count_for_epoch(epoch)
@@ -2211,1874 +2341,4 @@ impl TryFrom<Checkpoint> for ConsensusState {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::PublicKey;
-    use crate::account::{ValidatorAccount, ValidatorStatus};
-    use crate::execution_request::DepositRequest;
-    use crate::ssz_state_tree;
-    use crate::withdrawal::{PendingWithdrawal, WithdrawalKind};
-
-    use alloy_eips::eip4895::Withdrawal;
-    use alloy_primitives::Address;
-    use bytes::BytesMut;
-    use commonware_codec::{DecodeExt, Encode, ReadExt};
-    use commonware_consensus::types::{Epoch, Epocher, Height};
-    use commonware_cryptography::{Signer, bls12381, ed25519};
-
-    #[test]
-    fn test_read_truncated_input_returns_err() {
-        // Empty buffer — must not panic.
-        let empty: &[u8] = &[];
-        assert!(matches!(
-            ConsensusState::read(&mut empty.as_ref()),
-            Err(Error::EndOfBuffer)
-        ));
-
-        // Arbitrary short prefixes: each must return EndOfBuffer (not panic).
-        for n in 0..64 {
-            let data = vec![0xABu8; n];
-            let res = ConsensusState::read(&mut data.as_ref());
-            assert!(
-                res.is_err(),
-                "{n}-byte prefix should not successfully decode",
-            );
-        }
-    }
-
-    #[test]
-    fn test_decode_huge_deposit_queue_count_does_not_preallocate() {
-        // Regression guard for treating a length prefix as a safe element
-        // count. `deposit_queue_len` is an attacker-controlled u32; the decoder
-        // must reject a bogus count by running out of buffer, not by pre-sizing
-        // a VecDeque from it. (`buf.remaining()` is a byte count, so a
-        // count-derived capacity — even one capped by remaining bytes —
-        // over-allocates by size_of::<DepositRequest>() per slot.) With the
-        // count far exceeding the available bodies, decode must bail cheaply.
-        let mut buf = BytesMut::new();
-        buf.put_u64(0); // epoch
-        buf.put_u64(0); // view
-        buf.put_u64(0); // latest_height
-        buf.put_u32(u32::MAX); // claims ~4 billion deposits
-        // Provide a partial deposit body, then truncate. This leaves a non-zero
-        // `buf.remaining()` at the allocation point, so the original
-        // `with_capacity(len.min(buf.remaining()))` would have over-allocated
-        // `remaining`-many slots here rather than the degenerate zero.
-        buf.put_slice(&[0u8; 48]);
-
-        // DepositRequest::read_cfg runs out of buffer partway through the body;
-        // either way decode must fail cheaply rather than pre-allocate ~4
-        // billion slots.
-        let result = ConsensusState::read(&mut buf.as_ref());
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_decode_huge_captured_bytes_len_does_not_preallocate() {
-        // Regression guard for the captured-snapshot trailer: `captured_bytes`
-        // is a length-prefixed Vec<u8>, so a malformed blob with
-        // has_captured = true and a huge length must be rejected against the
-        // remaining bytes before the `vec![0u8; len]` allocation, not after.
-        let mut encoded = ConsensusState::default().encode().to_vec();
-        // A default state has captured_bytes = None, so the encoding ends with
-        // the has_captured presence flag (0). Flip it to 1 and append a u32
-        // length claiming ~4 GiB with no captured body following.
-        let last = encoded.len() - 1;
-        encoded[last] = 1;
-        encoded.extend_from_slice(&u32::MAX.to_be_bytes());
-
-        // Decode must bail cheaply on EndOfBuffer rather than pre-allocate ~4 GiB.
-        let result = ConsensusState::read(&mut encoded.as_slice());
-        assert!(result.is_err());
-    }
-
-    fn create_test_deposit_request(index: u64, amount: u64) -> DepositRequest {
-        let mut withdrawal_credentials = [0u8; 32];
-        withdrawal_credentials[0] = 0x01; // Eth1 withdrawal prefix
-        for i in 0..20 {
-            withdrawal_credentials[12 + i] = index as u8;
-        }
-
-        let consensus_key = bls12381::PrivateKey::from_seed(index);
-        DepositRequest {
-            node_pubkey: PublicKey::decode(&[1u8; 32][..]).unwrap(),
-            consensus_pubkey: consensus_key.public_key(),
-            withdrawal_credentials,
-            amount,
-            node_signature: [index as u8; 64],
-            consensus_signature: [index as u8; 96],
-            index,
-        }
-    }
-
-    fn create_test_withdrawal(index: u64, amount: u64, epoch: u64) -> PendingWithdrawal {
-        PendingWithdrawal {
-            inner: Withdrawal {
-                index,
-                validator_index: index * 10,
-                address: Address::from([index as u8; 20]),
-                amount,
-            },
-            pubkey: [index as u8; 32],
-            balance_deduction: amount,
-            epoch,
-            kind: WithdrawalKind::Validator,
-        }
-    }
-
-    fn create_test_validator_account(index: u64, balance: u64) -> ValidatorAccount {
-        let consensus_key = bls12381::PrivateKey::from_seed(1);
-        ValidatorAccount {
-            consensus_public_key: consensus_key.public_key(),
-            withdrawal_credentials: Address::from([index as u8; 20]),
-            balance,
-            status: ValidatorStatus::Active,
-            has_pending_deposit: false,
-            has_pending_withdrawal: false,
-            joining_epoch: 0,
-            last_deposit_index: index,
-        }
-    }
-
-    #[test]
-    fn test_serialization_deserialization_empty() {
-        let original_state = ConsensusState::default();
-
-        let mut encoded = original_state.encode();
-        let decoded_state = ConsensusState::decode(&mut encoded).expect("Failed to decode");
-
-        assert_eq!(decoded_state.epoch, original_state.epoch);
-        assert_eq!(decoded_state.view, original_state.view);
-        assert_eq!(decoded_state.latest_height, original_state.latest_height);
-        assert_eq!(
-            decoded_state.invalid_deposit_tax,
-            original_state.invalid_deposit_tax
-        );
-        assert_eq!(
-            decoded_state.get_next_withdrawal_index(),
-            original_state.get_next_withdrawal_index()
-        );
-        assert_eq!(
-            decoded_state.deposit_queue.len(),
-            original_state.deposit_queue.len()
-        );
-        assert_eq!(
-            decoded_state.withdrawal_queue,
-            original_state.withdrawal_queue
-        );
-        assert_eq!(
-            decoded_state.validator_accounts.len(),
-            original_state.validator_accounts.len()
-        );
-        assert_eq!(
-            decoded_state.epoch_genesis_hash,
-            original_state.epoch_genesis_hash
-        );
-        assert_eq!(
-            decoded_state.get_minimum_validator_count(),
-            DEFAULT_MINIMUM_VALIDATOR_COUNT
-        );
-        assert_eq!(decoded_state.get_pending_active_validator_exits(), 0);
-    }
-
-    #[test]
-    fn active_exit_counter_preserves_minimum_validator_count() {
-        let mut state = ConsensusState::default();
-        state.set_minimum_validator_count(3);
-
-        for i in 0..4 {
-            state.set_account(
-                [i as u8 + 1; 32],
-                create_test_validator_account(i as u64 + 1, 32_000_000_000),
-            );
-        }
-
-        assert!(state.can_accept_active_validator_exit());
-        state.increment_pending_active_validator_exits();
-        assert!(!state.can_accept_active_validator_exit());
-
-        let mut exiting_account = state.get_account(&[1u8; 32]).unwrap().clone();
-        exiting_account.status = ValidatorStatus::SubmittedExitRequest;
-        state.set_account([1u8; 32], exiting_account);
-        assert_eq!(state.current_epoch_active_validator_count(), 4);
-        assert!(!state.can_accept_active_validator_exit());
-
-        let refund = create_test_withdrawal(99, 1, 0);
-        state.push_withdrawal(refund);
-        assert_eq!(state.get_withdrawal_count_for_epoch(0), 1);
-        assert!(!state.can_accept_active_validator_exit());
-
-        state.reset_pending_active_validator_exits();
-        assert!(state.can_accept_active_validator_exit());
-    }
-
-    #[test]
-    fn exit_floor_honors_queued_minimum_validator_count_raise() {
-        // Removals staged this epoch take effect next epoch, at the same boundary
-        // a queued MinimumValidatorCount change applies — so the floor check must
-        // use the prospective value, not the current one.
-        let mut state = ConsensusState::default();
-        state.set_minimum_validator_count(2);
-        for i in 0..3u8 {
-            state.set_account(
-                [i + 1; 32],
-                create_test_validator_account(i as u64 + 1, 32_000_000_000),
-            );
-        }
-
-        // 3 active, floor 2: one exit is acceptable (3 - 1 >= 2).
-        assert!(state.can_accept_active_validator_exit());
-
-        // Queue a raise to floor 3. The prospective floor now governs: 3 - 1 = 2 < 3.
-        state.push_protocol_param_change(ProtocolParam::MinimumValidatorCount(3));
-        assert_eq!(state.prospective_minimum_validator_count(), 3);
-        assert!(!state.can_accept_active_validator_exit());
-
-        // A queued lowering is likewise honored before it is applied.
-        state.protocol_param_changes.clear();
-        state.push_protocol_param_change(ProtocolParam::MinimumValidatorCount(1));
-        assert_eq!(state.prospective_minimum_validator_count(), 1);
-        assert!(state.can_accept_active_validator_exit());
-    }
-
-    #[test]
-    fn test_clone_preserves_epoch_schedule_snapshot() {
-        let state = ConsensusState::new(
-            ForkchoiceState::default(),
-            0,
-            0,
-            NonZeroU64::new(10).unwrap(),
-            10_000,
-            Address::ZERO,
-            3,
-            16,
-            0,
-            0,
-            0,
-        );
-        state.get_epocher().advance_epoch(Epoch::new(0));
-
-        let cloned = state.clone();
-        let cloned_epoch_two_bounds_before = (
-            cloned.get_epocher().first(Epoch::new(2)),
-            cloned.get_epocher().last(Epoch::new(2)),
-        );
-
-        state
-            .get_epocher()
-            .update_length(NonZeroU64::new(20).unwrap())
-            .unwrap();
-        state.get_epocher().advance_epoch(Epoch::new(2));
-
-        assert_eq!(
-            (
-                cloned.get_epocher().first(Epoch::new(2)),
-                cloned.get_epocher().last(Epoch::new(2)),
-            ),
-            cloned_epoch_two_bounds_before,
-            "cloned consensus state must retain the epoch schedule captured at clone time",
-        );
-    }
-
-    #[test]
-    fn test_serialization_deserialization_populated() {
-        let mut original_state = ConsensusState::new(
-            ForkchoiceState::default(),
-            0,
-            0,
-            NonZeroU64::new(100).unwrap(),
-            10_000,
-            Address::ZERO,
-            3,
-            16,
-            0,
-            DEFAULT_MINIMUM_VALIDATOR_COUNT,
-            0,
-        );
-
-        original_state.set_epoch(7);
-        original_state.get_epocher().advance_epoch(Epoch::new(0));
-        original_state
-            .get_epocher()
-            .update_length(NonZeroU64::new(200).unwrap())
-            .unwrap();
-        original_state.get_epocher().advance_epoch(Epoch::new(7));
-        original_state.set_view(123);
-        original_state.set_latest_height(42);
-        original_state.set_next_withdrawal_index(5);
-        original_state.set_epoch_genesis_hash([42u8; 32]);
-        original_state.set_invalid_deposit_tax(25);
-
-        let deposit1 = create_test_deposit_request(1, 32000000000);
-        let deposit2 = create_test_deposit_request(2, 16000000000);
-        original_state.push_deposit(deposit1);
-        original_state.push_deposit(deposit2);
-
-        let withdrawal1 = create_test_withdrawal(1, 16000000000, 10);
-        let withdrawal2 = create_test_withdrawal(2, 24000000000, 11);
-        original_state.push_withdrawal(withdrawal1);
-        original_state.push_withdrawal(withdrawal2);
-
-        // Add protocol param changes
-        original_state.push_protocol_param_change(
-            crate::protocol_params::ProtocolParam::MinimumStake(40_000_000_000),
-        );
-        original_state.push_protocol_param_change(
-            crate::protocol_params::ProtocolParam::MaximumStake(80_000_000_000),
-        );
-        original_state
-            .push_protocol_param_change(crate::protocol_params::ProtocolParam::EpochLength(500));
-
-        let pubkey1 = [1u8; 32];
-        let pubkey2 = [2u8; 32];
-        let account1 = create_test_validator_account(1, 32000000000);
-        let account2 = create_test_validator_account(2, 64000000000);
-        original_state.set_account(pubkey1, account1);
-        original_state.set_account(pubkey2, account2);
-
-        // Add validators scheduled for future epochs
-        let validator1 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(10).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(10).public_key(),
-        };
-        let validator2 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(20).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(20).public_key(),
-        };
-        let validator3 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(30).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(30).public_key(),
-        };
-        let validator4 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(40).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(40).public_key(),
-        };
-
-        // Schedule validators for epoch 9 (current epoch + 2)
-        original_state.add_validator(9, validator1.clone());
-        original_state.add_validator(9, validator2.clone());
-
-        // Schedule validators for epoch 10
-        original_state.add_validator(10, validator3.clone());
-
-        // Schedule validators for epoch 11
-        original_state.add_validator(11, validator4.clone());
-
-        let mut encoded = original_state.encode();
-        let decoded_state = ConsensusState::decode(&mut encoded).expect("Failed to decode");
-
-        assert_eq!(decoded_state.epoch, original_state.epoch);
-        assert_eq!(decoded_state.view, original_state.view);
-        assert_eq!(decoded_state.latest_height, original_state.latest_height);
-        assert_eq!(
-            decoded_state.get_next_withdrawal_index(),
-            original_state.get_next_withdrawal_index()
-        );
-        assert_eq!(
-            decoded_state.epoch_genesis_hash,
-            original_state.epoch_genesis_hash
-        );
-
-        assert_eq!(decoded_state.deposit_queue.len(), 2);
-        assert_eq!(decoded_state.deposit_queue[0].amount, 32000000000);
-        assert_eq!(decoded_state.deposit_queue[1].amount, 16000000000);
-
-        // Check withdrawal_queue - two distinct scheduled epochs
-        assert_eq!(decoded_state.withdrawal_queue.num_epochs(), 2);
-
-        // At epoch 10, only the epoch-10 withdrawal is due (earliest epoch <= 10).
-        let epoch10_withdrawals = decoded_state.get_withdrawals_for_epoch(10);
-        assert_eq!(epoch10_withdrawals.len(), 1);
-        assert_eq!(epoch10_withdrawals[0].inner.index, 1);
-        assert_eq!(epoch10_withdrawals[0].inner.amount, 16000000000);
-
-        // By epoch 11 both are due (earliest epoch <= 11), in queue order.
-        let epoch11_withdrawals = decoded_state.get_withdrawals_for_epoch(11);
-        assert_eq!(epoch11_withdrawals.len(), 2);
-        assert_eq!(epoch11_withdrawals[1].inner.index, 2);
-        assert_eq!(epoch11_withdrawals[1].inner.amount, 24000000000);
-
-        // Verify protocol_param_changes
-        assert_eq!(decoded_state.protocol_param_changes.len(), 3);
-        match &decoded_state.protocol_param_changes[0] {
-            crate::protocol_params::ProtocolParam::MinimumStake(value) => {
-                assert_eq!(*value, 40_000_000_000)
-            }
-            _ => panic!("Expected MinimumStake variant"),
-        }
-        match &decoded_state.protocol_param_changes[1] {
-            crate::protocol_params::ProtocolParam::MaximumStake(value) => {
-                assert_eq!(*value, 80_000_000_000)
-            }
-            _ => panic!("Expected MaximumStake variant"),
-        }
-        match &decoded_state.protocol_param_changes[2] {
-            crate::protocol_params::ProtocolParam::EpochLength(value) => {
-                assert_eq!(*value, 500)
-            }
-            _ => panic!("Expected EpochLength variant"),
-        }
-
-        assert_eq!(decoded_state.validator_accounts.len(), 2);
-        let decoded_account1 = decoded_state.validator_accounts.get(&pubkey1).unwrap();
-        assert_eq!(decoded_account1.balance, 32000000000);
-        assert_eq!(decoded_account1.last_deposit_index, 1);
-        let decoded_account2 = decoded_state.validator_accounts.get(&pubkey2).unwrap();
-        assert_eq!(decoded_account2.balance, 64000000000);
-        assert_eq!(decoded_account2.last_deposit_index, 2);
-
-        // Verify added_validators
-        assert_eq!(decoded_state.added_validators.len(), 3);
-
-        // Check epoch 9 has 2 validators
-        let epoch9_validators = decoded_state.get_added_validators(9).unwrap();
-        assert_eq!(epoch9_validators.len(), 2);
-
-        // Check epoch 10 has 1 validator
-        let epoch10_validators = decoded_state.get_added_validators(10).unwrap();
-        assert_eq!(epoch10_validators.len(), 1);
-
-        // Check epoch 11 has 1 validator
-        let epoch11_validators = decoded_state.get_added_validators(11).unwrap();
-        assert_eq!(epoch11_validators.len(), 1);
-
-        // Check that epoch 8 returns None (no validators scheduled)
-        assert!(decoded_state.get_added_validators(8).is_none());
-
-        // Verify epocher round-trips correctly
-        let epocher = decoded_state.get_epocher();
-        assert_eq!(epocher.current_length(), 200);
-        // Epoch 0-1: length 100, epoch 2+: length 200
-        assert_eq!(epocher.first(Epoch::new(0)), Some(Height::new(0)));
-        assert_eq!(epocher.last(Epoch::new(1)), Some(Height::new(199)));
-        assert_eq!(epocher.first(Epoch::new(2)), Some(Height::new(200)));
-        assert_eq!(epocher.last(Epoch::new(2)), Some(Height::new(399)));
-    }
-
-    #[test]
-    fn test_encode_size_accuracy() {
-        let mut state = ConsensusState::default();
-
-        state.set_epoch(3);
-        state.set_view(456);
-        state.set_latest_height(42);
-        state.set_next_withdrawal_index(5);
-
-        let deposit = create_test_deposit_request(1, 32000000000);
-        state.push_deposit(deposit);
-
-        let withdrawal = create_test_withdrawal(1, 16000000000, 5);
-        state.push_withdrawal(withdrawal);
-
-        // Add protocol param changes
-        state.push_protocol_param_change(crate::protocol_params::ProtocolParam::MinimumStake(
-            50_000_000_000,
-        ));
-        state.push_protocol_param_change(crate::protocol_params::ProtocolParam::MaximumStake(
-            100_000_000_000,
-        ));
-
-        let pubkey = [1u8; 32];
-        let account = create_test_validator_account(1, 32000000000);
-        state.set_account(pubkey, account);
-
-        // Add validators scheduled for future epochs
-        let validator1 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(10).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(10).public_key(),
-        };
-        let validator2 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(20).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(20).public_key(),
-        };
-        let validator3 = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(30).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(30).public_key(),
-        };
-
-        state.add_validator(5, validator1.clone());
-        state.add_validator(6, validator2.clone());
-        state.add_validator(6, validator3.clone());
-
-        let predicted_size = state.encode_size();
-        let actual_encoded = state.encode();
-        let actual_size = actual_encoded.len();
-
-        assert_eq!(predicted_size, actual_size);
-    }
-
-    #[test]
-    fn pending_execution_requests_bind_into_captured_state_root() {
-        let mut state = ConsensusState::default();
-        state.rebuild_ssz_tree();
-        state.capture_state_root(0);
-        let before = state.get_state_root();
-
-        // Buffering a deferred request via the production mutator must change the
-        // captured state root (the mutator keeps the SSZ subtree in sync).
-        state.push_pending_execution_request(alloy_primitives::Bytes::from(vec![0xAAu8; 40]));
-        state.capture_state_root(0);
-        let after = state.get_state_root();
-        assert_ne!(
-            before, after,
-            "pushing a pending execution request must change the captured state root"
-        );
-
-        // Draining them restores the prior (empty-collection) root.
-        let taken = state.take_pending_execution_requests();
-        assert_eq!(taken.len(), 1);
-        state.capture_state_root(0);
-        assert_eq!(
-            state.get_state_root(),
-            before,
-            "draining pending requests must restore the prior state root"
-        );
-    }
-
-    #[test]
-    fn pending_checkpoint_binds_into_captured_state_root() {
-        let mut state = ConsensusState::default();
-        state.rebuild_ssz_tree();
-        state.capture_state_root(0);
-        let before = state.get_state_root();
-
-        // Setting the pending checkpoint via the production mutator binds its digest
-        // into the captured state root.
-        let checkpoint = Checkpoint::new(&state);
-        state.set_pending_checkpoint(Some(checkpoint));
-        state.capture_state_root(0);
-        let after = state.get_state_root();
-        assert_ne!(
-            before, after,
-            "setting a pending checkpoint must change the captured state root"
-        );
-
-        // Taking it restores the prior (no-checkpoint) root.
-        let taken = state.take_pending_checkpoint();
-        assert!(taken.is_some());
-        state.capture_state_root(0);
-        assert_eq!(
-            state.get_state_root(),
-            before,
-            "taking the pending checkpoint must restore the prior state root"
-        );
-    }
-
-    #[test]
-    fn dynamic_epoch_schedule_binds_into_captured_state_root() {
-        use std::num::NonZeroU64;
-
-        let mut state = ConsensusState::default();
-        state.rebuild_ssz_tree();
-        state.capture_state_root(0);
-        let before = state.get_state_root();
-
-        // Mutate the epoch schedule through interior mutability — no `&mut
-        // ConsensusState` setter is involved — and confirm the captured root still
-        // changes, via the refresh in `capture_state_root`.
-        state
-            .get_epocher()
-            .update_length(NonZeroU64::new(20).unwrap())
-            .expect("update_length should succeed");
-        state.capture_state_root(0);
-
-        assert_ne!(
-            before,
-            state.get_state_root(),
-            "an epoch-schedule change must change the captured state root"
-        );
-    }
-
-    /// Changing only a validator-account map key (the node
-    /// pubkey) must change the SSZ state root. The tree commits account values
-    /// positionally without the key, so two states with the same account value
-    /// under different keys must not share a root.
-    #[test]
-    fn validator_account_key_binds_into_state_root() {
-        let account = ValidatorAccount {
-            consensus_public_key: bls12381::PrivateKey::from_seed(1).public_key(),
-            withdrawal_credentials: Address::from([7u8; 20]),
-            balance: 32_000_000_000,
-            status: ValidatorStatus::Active,
-            has_pending_deposit: false,
-            has_pending_withdrawal: false,
-            joining_epoch: 0,
-            last_deposit_index: 0,
-        };
-
-        let root_for_key = |key: [u8; 32]| {
-            let mut state = ConsensusState::default();
-            state.validator_accounts.insert(key, account.clone());
-            state.rebuild_ssz_tree();
-            state.capture_state_root(0);
-            state.get_state_root()
-        };
-
-        assert_ne!(
-            root_for_key([1u8; 32]),
-            root_for_key([2u8; 32]),
-            "changing only the validator-account map key must change the state root"
-        );
-    }
-
-    /// Changing only the scheduled-activation epoch key must
-    /// change the SSZ state root. added_validators is flattened to its values, so
-    /// the same activation under a different epoch must not share a root.
-    #[test]
-    fn added_validator_epoch_key_binds_into_state_root() {
-        let av = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(1).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(1).public_key(),
-        };
-
-        let root_for_epoch = |epoch: u64| {
-            let mut state = ConsensusState::default();
-            state.add_validator(epoch, av.clone());
-            state.rebuild_ssz_tree();
-            state.capture_state_root(0);
-            state.get_state_root()
-        };
-
-        assert_ne!(
-            root_for_epoch(5),
-            root_for_epoch(6),
-            "changing only the added-validator epoch key must change the state root"
-        );
-    }
-
-    #[test]
-    fn test_protocol_param_changes_serialization() {
-        let mut state = ConsensusState::default();
-
-        // Add various protocol param changes
-        state.push_protocol_param_change(crate::protocol_params::ProtocolParam::MinimumStake(
-            32_000_000_000,
-        ));
-        state.push_protocol_param_change(crate::protocol_params::ProtocolParam::MaximumStake(
-            64_000_000_000,
-        ));
-        state.push_protocol_param_change(crate::protocol_params::ProtocolParam::MinimumStake(
-            40_000_000_000,
-        ));
-
-        let mut encoded = state.encode();
-        let decoded_state = ConsensusState::decode(&mut encoded).expect("Failed to decode");
-
-        assert_eq!(
-            decoded_state.protocol_param_changes.len(),
-            state.protocol_param_changes.len()
-        );
-        assert_eq!(decoded_state.protocol_param_changes.len(), 3);
-
-        match &decoded_state.protocol_param_changes[0] {
-            crate::protocol_params::ProtocolParam::MinimumStake(value) => {
-                assert_eq!(*value, 32_000_000_000)
-            }
-            _ => panic!("Expected MinimumStake variant"),
-        }
-
-        match &decoded_state.protocol_param_changes[1] {
-            crate::protocol_params::ProtocolParam::MaximumStake(value) => {
-                assert_eq!(*value, 64_000_000_000)
-            }
-            _ => panic!("Expected MaximumStake variant"),
-        }
-
-        match &decoded_state.protocol_param_changes[2] {
-            crate::protocol_params::ProtocolParam::MinimumStake(value) => {
-                assert_eq!(*value, 40_000_000_000)
-            }
-            _ => panic!("Expected MinimumStake variant"),
-        }
-
-        // Verify encode_size is correct
-        let predicted_size = state.encode_size();
-        let actual_size = state.encode().len();
-        assert_eq!(predicted_size, actual_size);
-    }
-
-    #[test]
-    fn test_decode_rejects_out_of_range_max_withdrawals_per_epoch() {
-        use crate::protocol_params::{
-            MAX_WITHDRAWALS_PER_EPOCH_MAX, MAX_WITHDRAWALS_PER_EPOCH_MIN,
-        };
-
-        // Honest nodes only ever serialize a cap within [MIN, MAX] — genesis and
-        // runtime updates both range-check it. A decoded state outside that range
-        // can only come from a crafted checkpoint/state artifact or a tampered DB
-        // blob. The finalizer trusts this cap as authoritative (a zero cap silently
-        // drops every due withdrawal), so decoding must reject it rather than let
-        // the node start/restore from it.
-
-        // Valid boundary values must still decode.
-        for valid in [MAX_WITHDRAWALS_PER_EPOCH_MIN, MAX_WITHDRAWALS_PER_EPOCH_MAX] {
-            let mut state = ConsensusState::default();
-            state.max_withdrawals_per_epoch = valid;
-            let encoded = state.encode();
-            let decoded = ConsensusState::read(&mut encoded.as_ref()).unwrap_or_else(|_| {
-                panic!("valid max_withdrawals_per_epoch {valid} should decode")
-            });
-            assert_eq!(decoded.max_withdrawals_per_epoch, valid);
-        }
-
-        // Out-of-range values (0 below MIN, MAX+1 above MAX) must be rejected.
-        for invalid in [0, MAX_WITHDRAWALS_PER_EPOCH_MAX + 1] {
-            let mut state = ConsensusState::default();
-            state.max_withdrawals_per_epoch = invalid;
-            let encoded = state.encode();
-            assert!(
-                ConsensusState::read(&mut encoded.as_ref()).is_err(),
-                "max_withdrawals_per_epoch {invalid} should be rejected on decode"
-            );
-        }
-    }
-
-    #[test]
-    fn test_decode_rejects_out_of_range_max_deposits_per_epoch() {
-        // Genesis and runtime updates cap deposits at MAX_MAX_DEPOSITS_PER_EPOCH;
-        // a decoded cap above it can only come from a crafted/tampered artifact and
-        // would let the penultimate-block selector admit more deposits than policy
-        // allows, so decode must reject it.
-        use crate::protocol_params::{MAX_MAX_DEPOSITS_PER_EPOCH, MIN_MAX_DEPOSITS_PER_EPOCH};
-
-        for valid in [MIN_MAX_DEPOSITS_PER_EPOCH, MAX_MAX_DEPOSITS_PER_EPOCH] {
-            let mut state = ConsensusState::default();
-            state.max_deposits_per_epoch = valid;
-            let encoded = state.encode();
-            let decoded = ConsensusState::read(&mut encoded.as_ref())
-                .unwrap_or_else(|_| panic!("valid max_deposits_per_epoch {valid} should decode"));
-            assert_eq!(decoded.max_deposits_per_epoch, valid);
-        }
-
-        let mut state = ConsensusState::default();
-        state.max_deposits_per_epoch = MAX_MAX_DEPOSITS_PER_EPOCH + 1;
-        let encoded = state.encode();
-        assert!(
-            ConsensusState::read(&mut encoded.as_ref()).is_err(),
-            "oversized max_deposits_per_epoch should be rejected on decode"
-        );
-    }
-
-    #[test]
-    fn test_decode_rejects_out_of_range_allowed_timestamp_future_ms() {
-        // The timestamp tolerance gates block-validity; genesis and runtime updates
-        // bound it to [MIN, MAX]. An out-of-range window from a crafted/tampered
-        // artifact would put a booting node's clock tolerance outside policy.
-        use crate::protocol_params::{
-            MAX_ALLOWED_TIMESTAMP_FUTURE_MS, MIN_ALLOWED_TIMESTAMP_FUTURE_MS,
-        };
-
-        for valid in [
-            MIN_ALLOWED_TIMESTAMP_FUTURE_MS,
-            MAX_ALLOWED_TIMESTAMP_FUTURE_MS,
-        ] {
-            let mut state = ConsensusState::default();
-            state.allowed_timestamp_future_ms = valid;
-            let encoded = state.encode();
-            let decoded = ConsensusState::read(&mut encoded.as_ref()).unwrap_or_else(|_| {
-                panic!("valid allowed_timestamp_future_ms {valid} should decode")
-            });
-            assert_eq!(decoded.allowed_timestamp_future_ms, valid);
-        }
-
-        for invalid in [
-            MIN_ALLOWED_TIMESTAMP_FUTURE_MS - 1,
-            MAX_ALLOWED_TIMESTAMP_FUTURE_MS + 1,
-        ] {
-            let mut state = ConsensusState::default();
-            state.allowed_timestamp_future_ms = invalid;
-            let encoded = state.encode();
-            assert!(
-                ConsensusState::read(&mut encoded.as_ref()).is_err(),
-                "allowed_timestamp_future_ms {invalid} should be rejected on decode"
-            );
-        }
-    }
-
-    #[test]
-    fn test_decode_rejects_out_of_range_observers_per_validator() {
-        // Genesis and runtime updates cap observers at MAX_OBSERVERS_PER_VALIDATOR;
-        // a decoded value above it can only come from a crafted/tampered artifact.
-        use crate::protocol_params::MAX_OBSERVERS_PER_VALIDATOR;
-
-        for valid in [0u32, MAX_OBSERVERS_PER_VALIDATOR as u32] {
-            let mut state = ConsensusState::default();
-            state.observers_per_validator = valid;
-            let encoded = state.encode();
-            let decoded = ConsensusState::read(&mut encoded.as_ref())
-                .unwrap_or_else(|_| panic!("valid observers_per_validator {valid} should decode"));
-            assert_eq!(decoded.observers_per_validator, valid);
-        }
-
-        let mut state = ConsensusState::default();
-        state.observers_per_validator = MAX_OBSERVERS_PER_VALIDATOR as u32 + 1;
-        let encoded = state.encode();
-        assert!(
-            ConsensusState::read(&mut encoded.as_ref()).is_err(),
-            "oversized observers_per_validator should be rejected on decode"
-        );
-    }
-
-    #[test]
-    fn test_account_operations() {
-        let mut state = ConsensusState::default();
-        let pubkey = [1u8; 32];
-        let account = create_test_validator_account(1, 32000000000);
-
-        // Test that account doesn't exist initially
-        assert!(state.get_account(&pubkey).is_none());
-
-        // Test setting account
-        state.set_account(pubkey, account.clone());
-        let retrieved_account = state.get_account(&pubkey);
-        assert!(retrieved_account.is_some());
-        assert_eq!(retrieved_account.unwrap().balance, account.balance);
-
-        // Test removing account
-        let removed_account = state.remove_account(&pubkey);
-        assert!(removed_account.is_some());
-        assert_eq!(removed_account.unwrap().balance, account.balance);
-
-        // Test that account no longer exists
-        assert!(state.get_account(&pubkey).is_none());
-
-        // Test removing non-existent account
-        let non_existent = state.remove_account(&pubkey);
-        assert!(non_existent.is_none());
-    }
-
-    #[test]
-    fn test_try_from_checkpoint() {
-        // Create a populated ConsensusState
-        let mut original_state = ConsensusState::default();
-        original_state.set_epoch(5);
-        original_state.set_view(789);
-        original_state.set_latest_height(100);
-        original_state.set_next_withdrawal_index(42);
-        original_state.set_epoch_genesis_hash([99u8; 32]);
-
-        // Add some data
-        let deposit = create_test_deposit_request(1, 32000000000);
-        original_state.push_deposit(deposit);
-
-        let withdrawal = create_test_withdrawal(1, 16000000000, 7);
-        original_state.push_withdrawal(withdrawal);
-
-        let pubkey = [1u8; 32];
-        let account = create_test_validator_account(1, 32000000000);
-        original_state.set_account(pubkey, account);
-
-        // Convert to checkpoint
-        let checkpoint = Checkpoint::new(&original_state);
-
-        // Convert back to ConsensusState
-        let restored_state: ConsensusState = checkpoint
-            .try_into()
-            .expect("Failed to convert checkpoint back to ConsensusState");
-
-        // Verify the data matches
-        assert_eq!(restored_state.epoch, original_state.epoch);
-        assert_eq!(restored_state.view, original_state.view);
-        assert_eq!(restored_state.latest_height, original_state.latest_height);
-        assert_eq!(
-            restored_state.get_next_withdrawal_index(),
-            original_state.get_next_withdrawal_index()
-        );
-        assert_eq!(
-            restored_state.epoch_genesis_hash,
-            original_state.epoch_genesis_hash
-        );
-        assert_eq!(
-            restored_state.deposit_queue.len(),
-            original_state.deposit_queue.len()
-        );
-        assert_eq!(
-            restored_state.withdrawal_queue,
-            original_state.withdrawal_queue
-        );
-        assert_eq!(
-            restored_state.validator_accounts.len(),
-            original_state.validator_accounts.len()
-        );
-
-        // Check specific values
-        assert_eq!(restored_state.deposit_queue[0].amount, 32000000000);
-        let epoch7_withdrawals = restored_state.get_withdrawals_for_epoch(7);
-        assert_eq!(epoch7_withdrawals[0].inner.amount, 16000000000);
-
-        let restored_account = restored_state.get_account(&pubkey).unwrap();
-        assert_eq!(restored_account.balance, 32000000000);
-        assert_eq!(restored_account.last_deposit_index, 1);
-    }
-
-    // ---- SSZ state tree integration tests ----
-
-    #[test]
-    fn test_ssz_scalar_setters_update_root() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-
-        state.set_epoch(10);
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        let r1 = state.ssz_tree().root();
-        state.set_view(99);
-        assert_ne!(state.ssz_tree().root(), r1);
-
-        let r2 = state.ssz_tree().root();
-        state.set_latest_height(500);
-        assert_ne!(state.ssz_tree().root(), r2);
-
-        let r3 = state.ssz_tree().root();
-        state.set_head_digest(sha256::Digest([0xAB; 32]));
-        assert_ne!(state.ssz_tree().root(), r3);
-
-        let r4 = state.ssz_tree().root();
-        state.set_epoch_genesis_hash([0xCD; 32]);
-        assert_ne!(state.ssz_tree().root(), r4);
-
-        let r5 = state.ssz_tree().root();
-        state.set_minimum_stake(16_000_000_000);
-        assert_ne!(state.ssz_tree().root(), r5);
-
-        let r6 = state.ssz_tree().root();
-        state.set_maximum_stake(64_000_000_000);
-        assert_ne!(state.ssz_tree().root(), r6);
-
-        let r7 = state.ssz_tree().root();
-        state.set_next_withdrawal_index(42);
-        assert_ne!(state.ssz_tree().root(), r7);
-    }
-
-    #[test]
-    fn test_ssz_scalar_proof_verifies() {
-        let mut state = ConsensusState::default();
-        state.set_epoch(10);
-        state.set_view(99);
-
-        let tree = state.ssz_tree();
-        let root = tree.root();
-        let proof = tree.generate_scalar_proof(ssz_state_tree::EPOCH);
-        assert!(proof.verify(&root));
-
-        let proof_view = tree.generate_scalar_proof(ssz_state_tree::VIEW);
-        assert!(proof_view.verify(&root));
-    }
-
-    #[test]
-    fn test_ssz_forkchoice_updates() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-
-        let fcs = ForkchoiceState {
-            head_block_hash: [0x11; 32].into(),
-            safe_block_hash: [0x22; 32].into(),
-            finalized_block_hash: [0x33; 32].into(),
-        };
-        state.set_forkchoice(fcs);
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        let r1 = state.ssz_tree().root();
-
-        // Partial setters
-        state.set_forkchoice_head([0xAA; 32].into());
-        assert_ne!(state.ssz_tree().root(), r1);
-
-        let r2 = state.ssz_tree().root();
-        state.set_forkchoice_safe_and_finalized([0xBB; 32].into());
-        assert_ne!(state.ssz_tree().root(), r2);
-    }
-
-    #[test]
-    fn test_ssz_validator_account_lifecycle() {
-        let mut state = ConsensusState::default();
-        let pubkey = [1u8; 32];
-        let account = create_test_validator_account(1, 32_000_000_000);
-
-        let root_before = state.ssz_tree().root();
-
-        // Insert
-        state.set_account(pubkey, account.clone());
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        // Verify proof
-        let tree = state.ssz_tree();
-        let root = tree.root();
-        let keys = [pubkey];
-        let proof = tree.generate_validator_proof(&pubkey, &keys).unwrap();
-        assert!(proof.verify(&root));
-
-        // Update balance
-        let mut updated = account.clone();
-        updated.balance = 48_000_000_000;
-        state.set_account(pubkey, updated);
-        assert_ne!(state.ssz_tree().root(), root);
-
-        // Remove
-        let root_with_account = state.ssz_tree().root();
-        state.remove_account(&pubkey);
-        assert_ne!(state.ssz_tree().root(), root_with_account);
-
-        // Validator proof should return None for removed pubkey
-        assert!(
-            state
-                .ssz_tree()
-                .generate_validator_proof(&pubkey, &[])
-                .is_none()
-        );
-    }
-
-    #[test]
-    fn test_ssz_deposit_queue_operations() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-
-        let deposit = create_test_deposit_request(1, 32_000_000_000);
-        state.push_deposit(deposit.clone());
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        let root_with_deposit = state.ssz_tree().root();
-
-        // Pop deposit changes root
-        let popped = state.pop_deposit().unwrap();
-        assert_eq!(popped.amount, 32_000_000_000);
-        assert_ne!(state.ssz_tree().root(), root_with_deposit);
-    }
-
-    #[test]
-    fn test_ssz_withdrawal_queue_operations() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-
-        let withdrawal = create_test_withdrawal(1, 16_000_000_000, 5);
-        state.push_withdrawal(withdrawal);
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        let root_with_withdrawal = state.ssz_tree().root();
-
-        // Pop withdrawal changes root
-        let popped = state.pop_withdrawal(5).unwrap();
-        assert_eq!(popped.inner.amount, 16_000_000_000);
-        assert_ne!(state.ssz_tree().root(), root_with_withdrawal);
-    }
-
-    #[test]
-    fn test_ssz_added_removed_validators() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-
-        let validator = AddedValidator {
-            node_key: ed25519::PrivateKey::from_seed(10).public_key(),
-            consensus_key: bls12381::PrivateKey::from_seed(10).public_key(),
-        };
-
-        // add_validator changes root
-        state.add_validator(5, validator.clone());
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        let root_with_added = state.ssz_tree().root();
-
-        // remove_added_validators_for_epoch changes root
-        state.remove_added_validators_for_epoch(5);
-        assert_ne!(state.ssz_tree().root(), root_with_added);
-
-        // push_removed_validator / clear_removed_validators
-        let removed_pk = ed25519::PrivateKey::from_seed(20).public_key();
-        let r1 = state.ssz_tree().root();
-        state.push_removed_validator(removed_pk);
-        assert_ne!(state.ssz_tree().root(), r1);
-
-        let r2 = state.ssz_tree().root();
-        state.clear_removed_validators();
-        assert_ne!(state.ssz_tree().root(), r2);
-    }
-
-    #[test]
-    fn test_ssz_protocol_param_changes() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-
-        state.push_protocol_param_change(ProtocolParam::MinimumStake(40_000_000_000));
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        let r1 = state.ssz_tree().root();
-        state.push_protocol_param_change(ProtocolParam::MaximumStake(80_000_000_000));
-        assert_ne!(state.ssz_tree().root(), r1);
-
-        // apply_protocol_parameter_changes consumes them
-        let changed = state.apply_protocol_parameter_changes().unwrap();
-        assert!(changed);
-        assert_eq!(state.get_minimum_stake(), 40_000_000_000);
-        assert_eq!(state.get_maximum_stake(), 80_000_000_000);
-
-        let root_before_tax = state.ssz_tree().root();
-        state.push_protocol_param_change(ProtocolParam::InvalidDepositTax(25));
-        assert_ne!(state.ssz_tree().root(), root_before_tax);
-        let changed = state.apply_protocol_parameter_changes().unwrap();
-        assert!(!changed);
-        assert_eq!(state.get_invalid_deposit_tax(), 25);
-
-        state.push_protocol_param_change(ProtocolParam::InvalidDepositTax(101));
-        let changed = state.apply_protocol_parameter_changes().unwrap();
-        assert!(!changed);
-        assert_eq!(state.get_invalid_deposit_tax(), 25);
-    }
-
-    #[test]
-    fn protocol_param_batch_accepts_valid_final_stake_interval() {
-        let mut state = ConsensusState::default();
-        state.push_protocol_param_change(ProtocolParam::MaximumStake(20_000_000_000));
-        state.push_protocol_param_change(ProtocolParam::MinimumStake(10_000_000_000));
-
-        let changed = state.apply_protocol_parameter_changes().unwrap();
-
-        assert!(changed);
-        assert_eq!(state.get_minimum_stake(), 10_000_000_000);
-        assert_eq!(state.get_maximum_stake(), 20_000_000_000);
-    }
-
-    #[test]
-    fn protocol_param_batch_rejects_inverted_final_stake_interval() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-        state.push_protocol_param_change(ProtocolParam::MinimumStake(80_000_000_000));
-
-        let err = state.apply_protocol_parameter_changes().unwrap_err();
-
-        assert!(matches!(err, Error::Invalid("ConsensusState", _)));
-        assert_eq!(state.get_minimum_stake(), 32_000_000_000);
-        assert_eq!(state.get_maximum_stake(), 32_000_000_000);
-        assert_eq!(state.ssz_tree().root(), root_before);
-        assert_eq!(state.protocol_param_changes.len(), 0);
-    }
-
-    #[test]
-    fn consensus_state_decode_rejects_inverted_stake_interval() {
-        let mut state = ConsensusState::default();
-        state.validator_minimum_stake = 80_000_000_000;
-        state.validator_maximum_stake = 32_000_000_000;
-
-        let mut encoded = state.encode();
-        let err = ConsensusState::decode(&mut encoded).unwrap_err();
-
-        assert!(matches!(err, Error::Invalid("ConsensusState", _)));
-    }
-
-    #[test]
-    fn test_ssz_rebuild_matches_incremental() {
-        let mut state = ConsensusState::default();
-
-        // Build up state incrementally through setters
-        state.set_epoch(7);
-        state.set_view(42);
-        state.set_latest_height(100);
-        state.set_head_digest(sha256::Digest([0xAB; 32]));
-        state.set_epoch_genesis_hash([0xCD; 32]);
-        state.set_minimum_stake(16_000_000_000);
-        state.set_maximum_stake(64_000_000_000);
-        state.set_next_withdrawal_index(5);
-        state.set_forkchoice(ForkchoiceState {
-            head_block_hash: [0x11; 32].into(),
-            safe_block_hash: [0x22; 32].into(),
-            finalized_block_hash: [0x33; 32].into(),
-        });
-
-        let pubkey = [1u8; 32];
-        state.set_account(pubkey, create_test_validator_account(1, 32_000_000_000));
-
-        let deposit = create_test_deposit_request(1, 32_000_000_000);
-        state.push_deposit(deposit);
-
-        let withdrawal = create_test_withdrawal(1, 16_000_000_000, 5);
-        state.push_withdrawal(withdrawal);
-
-        let incremental_root = state.ssz_tree().root();
-
-        // Rebuild from scratch
-        state.rebuild_ssz_tree();
-        let rebuilt_root = state.ssz_tree().root();
-
-        assert_eq!(incremental_root, rebuilt_root);
-    }
-
-    #[test]
-    fn test_ssz_root_survives_serialization_roundtrip() {
-        let mut state = ConsensusState::default();
-
-        state.set_epoch(5);
-        state.set_view(99);
-        state.set_latest_height(200);
-        state.set_next_withdrawal_index(10);
-        state.set_epoch_genesis_hash([0xFF; 32]);
-        state.set_forkchoice(ForkchoiceState {
-            head_block_hash: [0xAA; 32].into(),
-            safe_block_hash: [0xBB; 32].into(),
-            finalized_block_hash: [0xCC; 32].into(),
-        });
-
-        let pubkey = [1u8; 32];
-        state.set_account(pubkey, create_test_validator_account(1, 32_000_000_000));
-
-        let deposit = create_test_deposit_request(1, 32_000_000_000);
-        state.push_deposit(deposit);
-
-        let withdrawal = create_test_withdrawal(1, 16_000_000_000, 7);
-        state.push_withdrawal(withdrawal);
-
-        let original_root = state.ssz_tree().root();
-
-        // Round-trip through serialization
-        let mut encoded = state.encode();
-        let decoded = ConsensusState::decode(&mut encoded).unwrap();
-
-        assert_eq!(decoded.ssz_tree().root(), original_root);
-    }
-
-    #[test]
-    fn test_ssz_set_validator_accounts_rebuilds() {
-        let mut state = ConsensusState::default();
-        state.set_epoch(3);
-        state.set_account([1u8; 32], create_test_validator_account(1, 32_000_000_000));
-
-        let root_before = state.ssz_tree().root();
-
-        // Bulk replace validator accounts
-        let mut new_accounts = BTreeMap::new();
-        new_accounts.insert([2u8; 32], create_test_validator_account(2, 64_000_000_000));
-        new_accounts.insert([3u8; 32], create_test_validator_account(3, 48_000_000_000));
-        state.set_validator_accounts(new_accounts);
-
-        assert_ne!(state.ssz_tree().root(), root_before);
-
-        // New validators have proofs
-        let tree = state.ssz_tree();
-        let root = tree.root();
-        let keys = [[2u8; 32], [3u8; 32]];
-        let proof = tree.generate_validator_proof(&[2u8; 32], &keys).unwrap();
-        assert!(proof.verify(&root));
-
-        // Old validator is gone
-        assert!(tree.generate_validator_proof(&[1u8; 32], &keys).is_none());
-    }
-
-    #[test]
-    fn test_ssz_clone_independence() {
-        let mut state = ConsensusState::default();
-        state.set_epoch(5);
-        state.set_account([1u8; 32], create_test_validator_account(1, 32_000_000_000));
-
-        let cloned = state.clone();
-        let root_before = cloned.ssz_tree().root();
-
-        // Mutate original
-        state.set_epoch(99);
-        state.set_account([2u8; 32], create_test_validator_account(2, 64_000_000_000));
-
-        // Clone is unaffected
-        assert_eq!(cloned.ssz_tree().root(), root_before);
-    }
-
-    #[test]
-    fn test_ssz_capture_and_proof_tree() {
-        let mut state = ConsensusState::default();
-        state.set_epoch(5);
-        state.set_account([1u8; 32], create_test_validator_account(1, 32_000_000_000));
-
-        // Capture state root
-        state.capture_state_root(100);
-        let captured_root = state.get_state_root();
-        assert_eq!(captured_root, state.proof_tree().root());
-        assert_eq!(state.get_proof_el_block_number(), 100);
-
-        // Mutate the live tree
-        state.set_epoch(99);
-        assert_ne!(state.ssz_tree().root(), captured_root);
-
-        // Proof tree is still frozen at the captured state
-        assert_eq!(state.proof_tree().root(), captured_root);
-
-        // Proof still verifies against captured root
-        let proof = state
-            .proof_tree()
-            .generate_validator_proof(&[1u8; 32], state.proof_validator_keys())
-            .unwrap();
-        assert!(proof.verify(&captured_root));
-    }
-
-    /// A restart between `capture_state_root` and the next block must preserve
-    /// the captured snapshot: `state_root`, `proof_tree`, `proof_validator_keys`,
-    /// and `proof_el_block_number`. The finalizer captures the root inside
-    /// `execute_block` and only persists ConsensusState *after* the
-    /// epoch-transition mutations run, so the live SSZ tree at persistence
-    /// time differs from the captured one. If `Read` rebuilds the snapshot
-    /// from the post-mutation live fields, restarted validators end up with
-    /// a different aux-data `state_root` than uninterrupted peers — they
-    /// reject each other's proposals on `parent_beacon_block_root`.
-    #[test]
-    fn test_serialization_preserves_captured_proof_snapshot() {
-        // Build state with one validator and capture a snapshot.
-        let mut state = ConsensusState::default();
-        state.set_epoch(5);
-        state.set_account([1u8; 32], create_test_validator_account(1, 32_000_000_000));
-        state.capture_state_root(100);
-
-        let captured_root = state.get_state_root();
-        let captured_proof_root = state.proof_tree().root();
-        let captured_validator_keys = state.proof_validator_keys().to_vec();
-        let captured_el_block = state.get_proof_el_block_number();
-
-        // Mutate the live fields the same way an epoch-boundary apply does:
-        // bump epoch, swap a validator account out. Any live-tree mutation is
-        // sufficient — these specific ones ensure the post-mutation live root
-        // is provably different from the captured one.
-        state.set_epoch(99);
-        state.set_account([2u8; 32], create_test_validator_account(2, 32_000_000_000));
-        assert_ne!(
-            state.ssz_tree().root(),
-            captured_root,
-            "live tree mutations must produce a different root; the captured \
-             snapshot must NOT track them — this is the property the audit \
-             worries restart breaks"
-        );
-        // The frozen snapshot is unaffected by the live mutations: this is
-        // the invariant `capture_state_root` exists to provide, and it's the
-        // invariant the encode/decode roundtrip below must preserve.
-        assert_eq!(
-            state.get_state_root(),
-            captured_root,
-            "post-capture mutations must not touch the frozen state_root"
-        );
-
-        // Persist and restore.
-        let mut encoded = state.encode();
-        let restored = ConsensusState::decode(&mut encoded).expect("decode");
-
-        // Property 1: cross-validator block-validity agreement. A restarted
-        // validator and an uninterrupted peer both need to derive the same
-        // `parent_beacon_block_root` expectation; this is the field they use.
-        assert_eq!(
-            restored.get_state_root(),
-            captured_root,
-            "state_root must equal the pre-mutation captured root after restart, \
-             not the post-mutation live root"
-        );
-
-        // Property 2: proof generation. A restarted validator must be able to
-        // produce proofs that verify against the same on-chain root.
-        assert_eq!(
-            restored.proof_tree().root(),
-            captured_proof_root,
-            "proof_tree must reflect the captured snapshot, not the post-mutation tree"
-        );
-        assert_eq!(
-            restored.proof_validator_keys(),
-            captured_validator_keys.as_slice(),
-            "proof_validator_keys must be the captured snapshot"
-        );
-        assert_eq!(
-            restored.get_proof_el_block_number(),
-            captured_el_block,
-            "proof_el_block_number must be the captured value"
-        );
-
-        // End-to-end: a proof generated by the restored state must verify
-        // against the captured root.
-        let restored_proof = restored
-            .proof_tree()
-            .generate_validator_proof(&[1u8; 32], restored.proof_validator_keys())
-            .unwrap();
-        assert!(
-            restored_proof.verify(&captured_root),
-            "proof generated post-restart must verify against the captured root"
-        );
-    }
-
-    #[test]
-    fn test_ssz_push_withdrawal_request_keeps_next_index_in_sync() {
-        use crate::execution_request::WithdrawalRequest;
-
-        let mut state = ConsensusState::default();
-        state.set_epoch(1);
-        state.set_account([1u8; 32], create_test_validator_account(1, 32_000_000_000));
-
-        // push_withdrawal_request internally calls WithdrawalQueue::push_request
-        // which increments next_index. The SSZ tree's NEXT_WITHDRAWAL_INDEX leaf
-        // must stay in sync.
-        let request = WithdrawalRequest {
-            source_address: alloy_primitives::Address::from([0xAA; 20]),
-            validator_pubkey: [1u8; 32],
-            amount: 16_000_000_000,
-        };
-        state.push_withdrawal_request(request, 5, 16_000_000_000);
-
-        let incremental_root = state.ssz_tree().root();
-
-        // Rebuild must produce the same root
-        state.rebuild_ssz_tree();
-        let rebuilt_root = state.ssz_tree().root();
-
-        assert_eq!(
-            incremental_root, rebuilt_root,
-            "push_withdrawal_request must keep NEXT_WITHDRAWAL_INDEX in sync with rebuild"
-        );
-    }
-
-    /// Simulate the full block execution lifecycle and check that
-    /// incremental SSZ tree matches rebuild at every step.
-    #[test]
-    fn test_ssz_full_block_lifecycle_matches_rebuild() {
-        use crate::execution_request::WithdrawalRequest;
-        use crate::header::AddedValidator;
-        use crate::protocol_params::ProtocolParam;
-        use commonware_cryptography::Signer;
-
-        // Derive valid Ed25519 pubkeys from seeds
-        let ed_keys: Vec<ed25519::PrivateKey> = (1..=5u64)
-            .map(|i| ed25519::PrivateKey::from_seed(i))
-            .collect();
-        let pubkeys: Vec<[u8; 32]> = ed_keys
-            .iter()
-            .map(|k| k.public_key().as_ref().try_into().unwrap())
-            .collect();
-
-        // --- Genesis setup (mimics get_initial_state in args.rs) ---
-        let forkchoice = ForkchoiceState {
-            head_block_hash: [0xAA; 32].into(),
-            safe_block_hash: [0xAA; 32].into(),
-            finalized_block_hash: [0xAA; 32].into(),
-        };
-        let mut state = ConsensusState::new(
-            forkchoice,
-            32_000_000_000,
-            32_000_000_000,
-            NonZeroU64::new(10).unwrap(),
-            10_000,
-            Address::ZERO,
-            3,
-            16,
-            0,
-            DEFAULT_MINIMUM_VALIDATOR_COUNT,
-            0,
-        );
-
-        // Add 4 genesis validators (like the testnet)
-        for i in 0..4 {
-            state.set_account(
-                pubkeys[i],
-                create_test_validator_account(i as u64 + 1, 32_000_000_000),
-            );
-        }
-
-        // Check: after genesis setup, incremental matches rebuild
-        let genesis_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            genesis_root,
-            state.ssz_tree().root(),
-            "genesis: incremental != rebuild"
-        );
-
-        // --- Simulate execute_block for height 1 ---
-        state.set_forkchoice_head([0xBB; 32].into());
-        state.set_latest_height(1);
-        state.set_view(1);
-        state.set_head_digest([0xCC; 32].into());
-        state.capture_state_root(100);
-
-        let block1_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            block1_root,
-            state.ssz_tree().root(),
-            "block 1: incremental != rebuild"
-        );
-
-        // --- Simulate finalization (forkchoice update after capture) ---
-        state.set_forkchoice_safe_and_finalized([0xBB; 32].into());
-
-        let post_finalization_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            post_finalization_root,
-            state.ssz_tree().root(),
-            "post-finalization: incremental != rebuild"
-        );
-
-        // --- Simulate execute_block for height 2 (with a deposit) ---
-        state.set_forkchoice_head([0xDD; 32].into());
-
-        // Push a deposit request
-        let deposit = create_test_deposit_request(1, 32_000_000_000);
-        state.push_deposit(deposit);
-
-        state.set_latest_height(2);
-        state.set_view(2);
-        state.set_head_digest([0xEE; 32].into());
-        state.capture_state_root(101);
-
-        let block2_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            block2_root,
-            state.ssz_tree().root(),
-            "block 2: incremental != rebuild"
-        );
-
-        // --- Simulate execute_block for height 3 (pop deposit, push withdrawal) ---
-        state.set_forkchoice_head([0xFF; 32].into());
-
-        // Pop the deposit
-        let _ = state.pop_deposit();
-
-        // Process the deposit: create a new validator
-        let new_pubkey = pubkeys[4];
-        let mut new_account = create_test_validator_account(5, 32_000_000_000);
-        new_account.status = ValidatorStatus::Joining;
-        new_account.joining_epoch = 2;
-        state.set_account(new_pubkey, new_account);
-
-        // Add to added_validators
-        let node_key = ed_keys[4].public_key();
-        let consensus_key = bls12381::PrivateKey::from_seed(5).public_key();
-        state.add_validator(
-            2,
-            AddedValidator {
-                node_key,
-                consensus_key,
-            },
-        );
-
-        state.set_latest_height(3);
-        state.set_view(3);
-        state.set_head_digest([0x11; 32].into());
-        state.capture_state_root(102);
-
-        let block3_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            block3_root,
-            state.ssz_tree().root(),
-            "block 3: incremental != rebuild"
-        );
-
-        // --- Simulate epoch transition ---
-        // Apply protocol param changes (none in this case)
-        state.apply_protocol_parameter_changes().unwrap();
-
-        // Activate the joining validator
-        let mut account = state.get_account(&new_pubkey).unwrap().clone();
-        account.status = ValidatorStatus::Active;
-        state.set_account(new_pubkey, account);
-
-        // Clear added/removed validators
-        state.remove_added_validators_for_epoch(2);
-        state.clear_removed_validators();
-
-        // Increment epoch
-        state.set_epoch(2);
-        state.set_epoch_genesis_hash([0x22; 32]);
-
-        let epoch_transition_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            epoch_transition_root,
-            state.ssz_tree().root(),
-            "epoch transition: incremental != rebuild"
-        );
-
-        // --- Simulate withdrawal request ---
-        let wr = WithdrawalRequest {
-            source_address: alloy_primitives::Address::from([0xAA; 20]),
-            validator_pubkey: pubkeys[0],
-            amount: 32_000_000_000,
-        };
-        state.push_withdrawal_request(wr, 4, 32_000_000_000);
-
-        // Mark validator as exiting
-        let mut account = state.get_account(&pubkeys[0]).unwrap().clone();
-        account.balance = 0;
-        account.has_pending_withdrawal = true;
-        account.status = ValidatorStatus::Inactive;
-        state.set_account(pubkeys[0], account);
-
-        state.push_removed_validator(ed_keys[0].public_key());
-
-        let withdrawal_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            withdrawal_root,
-            state.ssz_tree().root(),
-            "withdrawal: incremental != rebuild"
-        );
-
-        // --- Simulate protocol param change ---
-        state.push_protocol_param_change(ProtocolParam::MinimumStake(16_000_000_000));
-        state.apply_protocol_parameter_changes().unwrap();
-
-        let param_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            param_root,
-            state.ssz_tree().root(),
-            "protocol param: incremental != rebuild"
-        );
-
-        // --- Remove validator account ---
-        state.remove_account(&pubkeys[0]);
-
-        let remove_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            remove_root,
-            state.ssz_tree().root(),
-            "remove validator: incremental != rebuild"
-        );
-    }
-
-    #[test]
-    fn test_withdrawal_requests_keep_ssz_tree_in_sync() {
-        let mut state = ConsensusState::default();
-
-        let req = |tag: u8, amount: u64| WithdrawalRequest {
-            source_address: Address::from([tag; 20]),
-            validator_pubkey: [tag; 32],
-            amount,
-        };
-
-        // Interleave validator withdrawals and deposit refunds. Pushing a validator
-        // withdrawal while a refund is already queued exercises the rebuild branch in
-        // `push_withdrawal_request_with_kind`; the rest are incremental appends.
-        state.push_withdrawal_request(req(1, 100), 5, 100);
-        state.push_refund_withdrawal_request(req(2, 200), 5, 0);
-        state.push_withdrawal_request(req(3, 300), 6, 300); // validator after a refund → rebuild
-        state.push_refund_withdrawal_request(req(4, 400), 7, 0);
-
-        // The incrementally maintained root must equal a full rebuild from the queue.
-        let incremental_root = state.ssz_tree().root();
-        state.rebuild_ssz_tree();
-        assert_eq!(
-            incremental_root,
-            state.ssz_tree().root(),
-            "incrementally maintained withdrawal SSZ root must match a full rebuild"
-        );
-    }
-
-    // A grouped batch of protocol param changes flushed
-    // through push_protocol_param_changes must land in exactly the same state
-    // (queue contents and ssz root) as pushing each record one at a time. the
-    // batch path rebuilds the param subtree once instead of once per record.
-    #[test]
-    fn test_batch_protocol_param_changes_match_per_record() {
-        use crate::protocol_params::ProtocolParam;
-
-        let params = vec![
-            ProtocolParam::MinimumStake(16_000_000_000),
-            ProtocolParam::MaximumStake(64_000_000_000),
-            ProtocolParam::EpochLength(128),
-            ProtocolParam::MaxDepositsPerEpoch(8),
-        ];
-
-        let mut per_record = ConsensusState::default();
-        for param in params.clone() {
-            per_record.push_protocol_param_change(param);
-        }
-
-        let mut batched = ConsensusState::default();
-        batched.push_protocol_param_changes(params.clone());
-
-        assert_eq!(
-            batched.protocol_param_changes.len(),
-            per_record.protocol_param_changes.len(),
-            "batched queue should match per record queue length"
-        );
-        assert_eq!(
-            batched.ssz_tree().root(),
-            per_record.ssz_tree().root(),
-            "batched ssz root should match per record root"
-        );
-
-        // the batch path is equivalent to a full rebuild from the same queue.
-        batched.rebuild_ssz_tree();
-        assert_eq!(
-            batched.ssz_tree().root(),
-            per_record.ssz_tree().root(),
-            "batched root should match a full rebuild"
-        );
-    }
-
-    // Genesis startup builds ConsensusState::new (which
-    // freezes the proof snapshot over an empty validator set), then inserts the
-    // genesis committee via set_account, which only touches the live tree. A
-    // rebuild_ssz_tree after materialization must re-freeze so the exposed
-    // state_root, proof_tree, and proof_validator_keys all commit to the
-    // installed committee, rather than staying stale until the first capture.
-    #[test]
-    fn test_genesis_materialization_refreshes_proof_snapshot() {
-        let mut state = ConsensusState::new(
-            ForkchoiceState::default(),
-            0,
-            0,
-            NonZeroU64::new(10).unwrap(),
-            10_000,
-            Address::ZERO,
-            3,
-            16,
-            0,
-            0,
-            0,
-        );
-
-        // mirror node/src/args.rs genesis materialization.
-        let mut keys: Vec<[u8; 32]> = Vec::new();
-        for i in 0..4u64 {
-            let mut pubkey = [0u8; 32];
-            pubkey[0] = i as u8 + 1;
-            state.set_account(pubkey, create_test_validator_account(i, 32_000_000_000));
-            keys.push(pubkey);
-        }
-        keys.sort();
-
-        // before re freezing, the frozen snapshot still reflects the empty set
-        // that new() captured, so it diverges from the live tree.
-        assert_ne!(
-            state.get_state_root(),
-            state.ssz_tree().root(),
-            "frozen root should be stale before the post genesis rebuild"
-        );
-
-        // the fix: re-freeze after the committee is installed.
-        state.rebuild_ssz_tree();
-
-        assert_eq!(
-            state.get_state_root(),
-            state.ssz_tree().root(),
-            "state_root should commit to the live tree after rebuild"
-        );
-        assert_eq!(
-            state.proof_tree().root(),
-            state.ssz_tree().root(),
-            "proof_tree should commit to the live tree after rebuild"
-        );
-        assert_eq!(
-            state.proof_validator_keys(),
-            keys.as_slice(),
-            "proof_validator_keys should list the genesis committee after rebuild"
-        );
-    }
-
-    // Draining a capped batch of deposits through
-    // pop_deposit_deferred + a single rebuild_deposit_tree must land in the
-    // exact same state (queue length and ssz root) as draining the same count
-    // one pop at a time, where every pop rebuilt the whole remaining subtree.
-    #[test]
-    fn test_deferred_deposit_drain_matches_per_pop() {
-        // backlog larger than the cap so the drain is partial and the remaining
-        // subtree is non trivial.
-        let backlog = 64usize;
-        let cap = 16usize;
-
-        let mut per_pop = ConsensusState::default();
-        let mut deferred = ConsensusState::default();
-        for i in 0..backlog as u64 {
-            let deposit = create_test_deposit_request(i, 32_000_000_000 + i);
-            per_pop.push_deposit(deposit.clone());
-            deferred.push_deposit(deposit);
-        }
-        assert_eq!(
-            per_pop.ssz_tree().root(),
-            deferred.ssz_tree().root(),
-            "states should start identical"
-        );
-
-        // per pop path: rebuild on every pop (the original behaviour).
-        for _ in 0..cap {
-            per_pop.pop_deposit();
-        }
-
-        // deferred path: pop without rebuilding, then rebuild exactly once.
-        for _ in 0..cap {
-            deferred.pop_deposit_deferred();
-        }
-        deferred.rebuild_deposit_tree();
-
-        assert_eq!(
-            deferred.deposit_count(),
-            per_pop.deposit_count(),
-            "both paths should drain the same number of deposits"
-        );
-        assert_eq!(deferred.deposit_count(), backlog - cap);
-        assert_eq!(
-            deferred.ssz_tree().root(),
-            per_pop.ssz_tree().root(),
-            "deferred single rebuild root should match per pop root"
-        );
-
-        // and the deferred root must equal a fresh full rebuild from the queue.
-        deferred.rebuild_ssz_tree();
-        assert_eq!(
-            deferred.ssz_tree().root(),
-            per_pop.ssz_tree().root(),
-            "deferred root should match a full rebuild"
-        );
-    }
-
-    // draining a backlog smaller than the cap must fully empty the queue and
-    // leave a root identical to per pop draining (mirrors the finalizer break
-    // on empty queue).
-    #[test]
-    fn test_deferred_deposit_drain_empties_small_backlog() {
-        let backlog = 5usize;
-        let cap = 16usize;
-
-        let mut per_pop = ConsensusState::default();
-        let mut deferred = ConsensusState::default();
-        for i in 0..backlog as u64 {
-            let deposit = create_test_deposit_request(i, 32_000_000_000 + i);
-            per_pop.push_deposit(deposit.clone());
-            deferred.push_deposit(deposit);
-        }
-
-        let mut drained_any = false;
-        for _ in 0..cap {
-            if per_pop.pop_deposit().is_none() {
-                break;
-            }
-        }
-        for _ in 0..cap {
-            if deferred.pop_deposit_deferred().is_some() {
-                drained_any = true;
-            } else {
-                break;
-            }
-        }
-        if drained_any {
-            deferred.rebuild_deposit_tree();
-        }
-
-        assert_eq!(deferred.deposit_count(), 0);
-        assert_eq!(per_pop.deposit_count(), 0);
-        assert_eq!(
-            deferred.ssz_tree().root(),
-            per_pop.ssz_tree().root(),
-            "empty queue roots should match"
-        );
-    }
-
-    // an empty batch must be a no op: no queue growth and no root change, so the
-    // finalizer can call it unconditionally without forcing a needless rebuild.
-    #[test]
-    fn test_empty_protocol_param_batch_is_noop() {
-        let mut state = ConsensusState::default();
-        let root_before = state.ssz_tree().root();
-        let len_before = state.protocol_param_changes.len();
-
-        state.push_protocol_param_changes(std::iter::empty());
-
-        assert_eq!(len_before, state.protocol_param_changes.len());
-        assert_eq!(
-            root_before,
-            state.ssz_tree().root(),
-            "empty batch should not change the ssz root"
-        );
-    }
-}
+mod tests;
